Remove any paths from excluded analysis
Changes in 2.10.0 cause any analysis, including errors, to be
ignored if you exclude a file in `analysis_options`. Removing this
as a forced option from `resideo_pedantic` because there isn't
a way to override this if you inherit analysis options.

If needed, add the ignored directories directly to your package/app.

```dart

analyzer:
  exclude:
    - test_driver/**
    - test/**
    - "**/*.g.dart"
    - "**/*.freezed.dart"

```